pydynamo/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pydynamo"
version = "0.1.0"
description = "Python CLI for Dynamo subtomogram averaging (crop, reconstruction, alignment, MRA)"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "mrcfile>=1.0",
    "starfile>=0.4",
    "pandas>=1.3",
    "pyyaml>=6.0",
    "torch>=1.9",
    "eulerangles>=0.2",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[project.scripts]
pydynamo = "pydynamo.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
